refactor CutOut layer
