+++ examples/layers/preprocessing/cut_out_demo.py
@@ -6,7 +6,7 @@ Finally, they are shown using matplotlib.
 """
 import tensorflow as tf
 import tensorflow_datasets as tfds
-from keras_cv.layers.preprocessing import erasing
+from keras_cv.layers.preprocessing import cut_out
 import matplotlib.pyplot as plt
 
 
@@ -31,7 +31,7 @@ def main():
         .shuffle(10 * BATCH_SIZE)
         .batch(BATCH_SIZE)
     )
-    cutout = erasing.CutOut(1.0, length=50, patch_value=0.0)
+    cutout = cut_out.CutOut(1.0, length=50, fill_value=0.0)
     train_ds = train_ds.map(cutout, num_parallel_calls=tf.data.AUTOTUNE)
 
     for images, labels in train_ds.take(1):
